% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{IC_SP_Waldcc}
\alias{IC_SP_Waldcc}
\title{Confidence Interval for Seroprevalence with Continuity Correction (Wald Method)}
\usage{
IC_SP_Waldcc(SP, n, conf.level = 0.95)
}
\arguments{
\item{SP}{Seroprevalence estimate.}

\item{n}{Sample size.}

\item{conf.level}{Confidence level (default is 0.95).}
}
\value{
A vector with the lower and upper limits of the confidence interval.
}
\description{
Calculates the confidence interval for seroprevalence using the Wald method with continuity correction.
}
\examples{
IC_SP_Waldcc(0.25, 100, conf.level = 0.95)
}
