% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence.R
\name{prob_seropositive}
\alias{prob_seropositive}
\title{Calculation of Seropositivity Probability}
\usage{
prob_seropositive(SCR, SRR, t)
}
\arguments{
\item{SCR}{Seroconversion Rate}

\item{SRR}{Seroreversion Rate.}

\item{t}{Age for which we want to calculate the probability of seropositivity.}
}
\value{
The probability of seropositivity for age `t`.
}
\description{
This function calculates the probability of seropositivity based on the age and the seroconversion and seroreversion rates,
 using a reversible catalytic model.
}
\examples{
prob_seropositive(0.03, 0.01, 45)
}
\references{
For more information on the reversible catalytic model, see
\url{https://link.springer.com/article/10.1186/s12936-015-0661-z}
}
