% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMsize-package.R
\docType{package}
\name{RCMsize-package}
\alias{RCMsize}
\alias{RCMsize-package}
\title{RCMsize: Sample Size Calculation in Reversible Catalytic Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Sample size and confidence interval calculations in reversible catalytic models, with applications in malaria research. Further details can be found in the paper by Sepúlveda and Drakeley (2015, \doi{10.1186/s12936-015-0661-z}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marciagraca/RCMsize }
  \item \url{https://marciagraca.github.io/RCMsize/}
  \item Report bugs at \url{https://github.com/marciagraca/RCMsize/issues}
}

}
\author{
\strong{Maintainer}: Márcia Graça \email{marcia_da_graca@hotmail.com} [copyright holder]

}
\keyword{internal}
